+++ src/test/java/com/assignmentTest/Assignment6Test.java
@@ -1,0 +1,12 @@
+package com.assignmentTest;
+import org.junit.jupiter.api.Test;
+import com.assignment.Assignment6;
+
+public class Assignment6Test {
+
+    @Test
+    public void printCompanyReview() {
+     Assignment6 Test=new Assignment6();
+     System.out.println(Test.companyReview("Google"));
+     }
+}
